requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
pandas
google-api-python-client
jinja2
google-oauth2-tool
